fix 1.19
Traceback (most recent call last):
  File "/home/zx/data/Games/mc-1.18.1/./MCLauncher.pyz/launcher.py", line 189, in clear_natives
AttributeError: 'MCL' object has no attribute 'natives_dll_path'

+++ launcher.py
@@ -186,8 +186,9 @@ class MCL:
                     zf.extract(name, target)
 
     def clear_natives(self):
-        logger.debug(f"清理native库: {self.natives_dll_path}")
-        rmtree(self.natives_dll_path)
+        if hasattr(self, "natives_dll_path"):
+            logger.debug(f"清理native库: {self.natives_dll_path}")
+            rmtree(self.natives_dll_path)
 
     def get_classpath(self):
